small tweaks after merge

+++ Api/Client/Mappers/ClientToUi.cs
@@ -27,7 +27,7 @@
             return new()
             {
                 Id = (int)movie.Id,
-                ImdbId = (int)movie.ImdbId,
+                ImdbId = movie.ImdbId,
                 PosterPath = movie.PosterPath,
                 Title = movie.Title,
                 OriginalTitle = movie.OriginalTitle,
